requirements.txt
beautifulsoup4
google-generativeai
fpdf2
requests
